Divide inpainting class
