+++ Assets/AXIP/AILIA-MODELS/Diffusion/AiliaDiffusionInpainting.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d09fba70ca278491a9aa9ca86abab3c8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
